+++ Code/Unit.h
@@ -84,7 +84,7 @@ public:
 	 *
 	 * @return double
 	 */
-	virtual double getAverageMoral();
+	virtual int getAverageMoral();
 
 	/**
 	 * @brief Returns the current damage of the unit
